refactor(playback): split pure loop-utility statics into GhostPlaybackEngine.LoopBounds.cs
